Use PKCS7 padding conventions.
This makes SimpleAES compatible with many of the other AES
implementations out there, most notably CryptoJS.  It uses PKCS7-style
padding for the byte stream, which is a revertible, unambiguous padding
that avoids data loss while not requiring length announcements in the
stream header.

This is the default pad-encoding for CryptoJS and possibly many more
algorithms out there.